requirements.txt
Django~=2.1.2
djangorestframework~=3.8.2
